swap in the correctly mapping RGBW color.py
