+++ shows/cycle_hsv.py
@@ -67,6 +67,8 @@ class CycleHSV(ShowBase):
                 print('CB', self.cb.hsv)
 
             self.tri_grid.clear()
+            self.tri_grid.set_all_cells(RGBW(0,0,0,255))
             time.sleep(3)
+            self.tri_grid.clear()
 
             yield self.frame_delay
